requirements.txt
Flask==2.3.2
Flask==2.0.1
Flask_CKEditor==0.5.0
Flask_SQLAlchemy==2.5.0
Flask_WTF==1.1.1
PyJWT==2.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==2.3.6
Werkzeug==2.0.2
WTForms==3.0.1
